chore: cleanup and reorg methods
chore: remove panics add Result

chore: cargo lock
