+++ bitvm/src/chunk/helpers.rs
@@ -1,21 +1,8 @@
 use ark_ff::{BigInt, BigInteger};
 
 use crate::bigint::U256;
-use crate::bn254::fq2::Fq2;
-use crate::chunk::wrap_hasher::{hash_n_bytes, BLAKE3_HASH_LENGTH};
-use crate::signatures::wots_api::{wots160, wots256};
-use crate::{
-    bn254::{fp254impl::Fp254Impl, fq::Fq},
-    treepp::*,
-};
-
-pub(crate) type HashBytes = [u8; 64];
-
-#[derive(Debug, Clone)]
-pub enum SigData {
-    Sig256(wots256::Signature),
-    Sig160(wots160::Signature),
-}
+use crate::chunk::wrap_hasher::BLAKE3_HASH_LENGTH;
+use crate::treepp::*;
 
 #[cfg(test)]
 pub(crate) fn unpack_limbs_to_nibbles() -> Script {
@@ -28,23 +15,6 @@ pub fn pack_nibbles_to_limbs() -> Script {
 
 pub fn pack_bytes_to_limbs() -> Script {
     U256::transform_limbsize(8,29)
-}
-
-
-pub(crate) fn hash_fp2() -> Script {
-    script! {
-        { hash_n_bytes::<64>() }
-        { pack_nibbles_to_limbs() }
-    }
-}
-
-pub(crate) fn hash_fp4() -> Script {
-    script! {
-        // [a0b0, a0b1, a1b0, a1b1]
-        {Fq2::roll(2)}
-        { hash_n_bytes::<128>() }
-        { pack_nibbles_to_limbs() }
-    }
 }
 
 pub(crate) fn extern_bigint_to_nibbles(msg: ark_ff::BigInt<4>) -> [u8; 64] {
@@ -67,6 +37,75 @@ fn fq_to_chunked_bits(fq: BigInt<4>, limb_size: usize) -> Vec<u32> {
                 res
         })
         .collect()
+}
+
+pub(crate) fn extern_hash_fps(fqs: Vec<ark_bn254::Fq>) -> [u8; 64] {
+    let mut msgs: Vec<[u8; 64]> = Vec::new();
+    for fq in fqs {
+        let v = fq_to_chunked_bits(fq.into(), 4);
+        let nib_arr: Vec<u8> = v.into_iter().map(|x| x as u8).collect();
+        msgs.push(nib_arr.try_into().unwrap());
+    }
+    extern_hash_nibbles(msgs)
+}
+
+pub(crate) fn extern_hash_nibbles(msgs: Vec<[u8; 64]>) -> [u8; 64] {
+    assert!(msgs.len() == 4 || msgs.len() == 2 || msgs.len() == 12 || msgs.len() == 14 || msgs.len() == 6 || msgs.len() == 8);
+
+    fn hex_string_to_nibble_array(hex_string: &str) -> Vec<u8> {
+        hex_string
+            .chars()
+            .map(|c| c.to_digit(16).expect("Invalid hex character") as u8) // Convert each char to a nibble
+            .collect()
+    }
+
+    fn nib_to_byte_array(digits: &[u8]) -> Vec<u8> {
+        let mut msg_bytes = Vec::with_capacity(digits.len() / 2);
+    
+        for nibble_pair in digits.chunks(2) {
+            let byte = (nibble_pair[0] << 4) | (nibble_pair[1] & 0b00001111);
+            msg_bytes.push(byte);
+        }
+    
+        fn le_to_be_byte_array(byte_array: Vec<u8>) -> Vec<u8> {
+            assert!(byte_array.len() % 4 == 0, "Byte array length must be a multiple of 4");
+            byte_array
+                .chunks(4) // Process each group of 4 bytes (one u32)
+                .flat_map(|chunk| chunk.iter().rev().cloned()) // Reverse each chunk
+                .collect()
+        }
+        le_to_be_byte_array(msg_bytes)
+    }
+
+    fn replace_first_n_with_zero(hex_string: &str, n: usize) -> String {
+        let mut result = String::new();
+    
+        if hex_string.len() <= n {
+            result.push_str(&"0".repeat(hex_string.len())); // If n >= string length, replace all
+        } else {
+            result.push_str(&"0".repeat(n)); // Replace first n characters
+            result.push_str(&hex_string[0..(hex_string.len()-n)]); // Keep the rest of the string
+        }
+        result
+    }
+
+    fn extern_hash_fp_var(fqs: Vec<[u8; 64]>) -> [u8;64] {
+        let mut vs = Vec::new();
+        for fq in fqs {
+            let v = fq.to_vec();
+            vs.extend_from_slice(&v);
+        }
+        let nib_arr: Vec<u8> = vs.clone().into_iter().collect();
+        let p_bytes:Vec<u8> = nib_to_byte_array(&nib_arr);
+
+        let hash_out = blake3::hash(&p_bytes).to_string();
+
+        let hash_out = replace_first_n_with_zero(&hash_out.to_string(), (32-BLAKE3_HASH_LENGTH)*2);
+        let res = hex_string_to_nibble_array(&hash_out);
+        res.try_into().unwrap()
+    }
+
+    extern_hash_fp_var(msgs)
 }
 
 pub(crate) fn extern_nibbles_to_bigint(nibble_array: [u8; 64]) -> ark_ff::BigInt<4> {
@@ -108,139 +147,6 @@ pub(crate) fn extern_nibbles_to_limbs(nibble_array: [u8; 64]) -> [u32; 9] {
     limbs.try_into().unwrap()
 }
 
-
-fn replace_first_n_with_zero(hex_string: &str, n: usize) -> String {
-    let mut result = String::new();
-
-    if hex_string.len() <= n {
-        result.push_str(&"0".repeat(hex_string.len())); // If n >= string length, replace all
-    } else {
-        result.push_str(&"0".repeat(n)); // Replace first n characters
-        result.push_str(&hex_string[0..(hex_string.len()-n)]); // Keep the rest of the string
-    }
-    result
-}
-
-pub(crate) fn extern_hash_fps(fqs: Vec<ark_bn254::Fq>) -> [u8; 64] {
-    let mut msgs: Vec<[u8; 64]> = Vec::new();
-    for fq in fqs {
-        let v = fq_to_chunked_bits(fq.into(), 4);
-        let nib_arr: Vec<u8> = v.into_iter().map(|x| x as u8).collect();
-        msgs.push(nib_arr.try_into().unwrap());
-    }
-    extern_hash_nibbles(msgs)
-}
-
-pub(crate) fn extern_hash_nibbles(msgs: Vec<[u8; 64]>) -> [u8; 64] {
-    assert!(msgs.len() == 4 || msgs.len() == 2 || msgs.len() == 12 || msgs.len() == 14 || msgs.len() == 6 || msgs.len() == 8);
-
-    fn hex_string_to_nibble_array(hex_string: &str) -> Vec<u8> {
-        hex_string
-            .chars()
-            .map(|c| c.to_digit(16).expect("Invalid hex character") as u8) // Convert each char to a nibble
-            .collect()
-    }
-
-    fn nib_to_byte_array(digits: &[u8]) -> Vec<u8> {
-        let mut msg_bytes = Vec::with_capacity(digits.len() / 2);
-    
-        for nibble_pair in digits.chunks(2) {
-            let byte = (nibble_pair[0] << 4) | (nibble_pair[1] & 0b00001111);
-            msg_bytes.push(byte);
-        }
-    
-        fn le_to_be_byte_array(byte_array: Vec<u8>) -> Vec<u8> {
-            assert!(byte_array.len() % 4 == 0, "Byte array length must be a multiple of 4");
-            byte_array
-                .chunks(4) // Process each group of 4 bytes (one u32)
-                .flat_map(|chunk| chunk.iter().rev().cloned()) // Reverse each chunk
-                .collect()
-        }
-        le_to_be_byte_array(msg_bytes)
-    }
-
-    fn extern_hash_fp_var(fqs: Vec<[u8; 64]>) -> [u8;64] {
-        let mut vs = Vec::new();
-        for fq in fqs {
-            let v = fq.to_vec();
-            vs.extend_from_slice(&v);
-        }
-        let nib_arr: Vec<u8> = vs.clone().into_iter().collect();
-        let p_bytes:Vec<u8> = nib_to_byte_array(&nib_arr);
-
-        let hash_out = blake3::hash(&p_bytes).to_string();
-
-        let hash_out = replace_first_n_with_zero(&hash_out.to_string(), (32-BLAKE3_HASH_LENGTH)*2);
-        let res = hex_string_to_nibble_array(&hash_out);
-        res.try_into().unwrap()
-    }
-
-    extern_hash_fp_var(msgs)
-}
-
-pub(crate) fn new_hash_g2acc_with_hashed_le() -> Script {
-    script! {
-        //Stack: [tx, ty, hash_inaux]
-        //T
-        {Fq::toaltstack()} 
-        {hash_fp4()} // HT
-
-        {Fq::fromaltstack()}
-        {hash_fp2()}
-    }
-}
-
-pub(crate) fn new_hash_g2acc() -> Script {
-    script!{
-        // [t, le]
-        for _ in 0..14 {
-            {Fq::toaltstack()}
-        }
-        {hash_fp4()}
-        for _ in 0..14 {
-            {Fq::fromaltstack()}
-        }
-        {Fq::roll(14)} {Fq::toaltstack()}
-        {hash_fp14()}
-        {Fq::fromaltstack()}
-        {Fq::roll(1)}
-        {hash_fp2()}
-    }
-}
-
-pub(crate) fn new_hash_g2acc_with_hash_t() -> Script {
-    script!{
-        // [le, ht]
-        {Fq::toaltstack()}
-        {hash_fp14()}
-        {Fq::fromaltstack()}
-        {Fq::roll(1)}
-        {hash_fp2()}
-    }
-}
-
-
-pub(crate) fn hash_fp6() -> Script {
-    script! {
-        {Fq2::roll(2)} {Fq2::roll(4)}
-        {hash_n_bytes::<192>()}
-        {pack_nibbles_to_limbs()}
-
-    }
-}
-
-
-pub(crate) fn hash_fp14() -> Script {
-    script! {
-        {Fq2::roll(2)} {Fq2::roll(4)} {Fq2::roll(6)} 
-        {Fq2::roll(8)} {Fq2::roll(10)} {Fq2::roll(12)}
-        {hash_n_bytes::<448>()}
-        {pack_nibbles_to_limbs()}
-
-    }
-}
-
-
 #[cfg(test)]
 mod test {
     use super::*;
@@ -248,7 +154,8 @@ mod test {
     use rand::SeedableRng;
     use rand_chacha::ChaCha20Rng;
 
-    use crate::execute_script;
+    use crate::{chunk::wrap_hasher::hash_utils::{hash_fp14, hash_fp2, hash_fp4, hash_fp6}, execute_script};
+    use crate::bn254::fq::Fq;
 
     #[test]
     fn test_emulate_fq_to_nibbles() {
