+++ bitvm/src/chunk/mod.rs
@@ -1,7 +1,7 @@
 mod wrap_hasher;
 mod wrap_wots;
 
-mod primitives;
+mod helpers;
 mod elements;
 
 pub mod api;
@@ -10,8 +10,6 @@ pub mod api_compiletime_utils;
 
 pub mod g16_runner_core;
 pub mod g16_runner_utils;
-
-pub mod assigner;
 
 mod taps_msm;
 mod taps_point_ops;
