+++ bitvm/src/chunk/wrap_wots.rs
@@ -1,6 +1,6 @@
 use bitcoin_script::script;
 
-use crate::chunk::primitives::pack_bytes_to_limbs;
+use crate::chunk::helpers::pack_bytes_to_limbs;
 use crate::chunk::wrap_hasher::BLAKE3_HASH_LENGTH;
 use crate::pseudo::NMUL;
 use crate::treepp::Script;
@@ -43,13 +43,11 @@ pub(crate) fn checksig_verify_to_limbs(pub_key: &WOTSPubKey) -> Script {
     }
 }
 
-#[cfg(test)]
-fn byte_array_to_wots160_sig(secret: &str, msg_bytes: &[u8]) -> wots160::Signature {
+pub(crate) fn byte_array_to_wots160_sig(secret: &str, msg_bytes: &[u8]) -> wots160::Signature {
     wots160::get_signature(secret, msg_bytes)
 } 
 
-#[cfg(test)]
-fn byte_array_to_wots256_sig(secret: &str, msg_bytes: &[u8]) -> wots256::Signature {
+pub(crate) fn byte_array_to_wots256_sig(secret: &str, msg_bytes: &[u8]) -> wots256::Signature {
     wots256::get_signature(secret, msg_bytes)
 } 
 
@@ -94,7 +92,7 @@ mod test {
     use rand::SeedableRng;
     use rand_chacha::ChaCha20Rng;
     use bitcoin_script::script;
-    use crate::{bn254::{fp254impl::Fp254Impl, fq::Fq}, chunk::{elements::CompressedStateObject, primitives::extern_hash_fps, wrap_wots::{byte_array_to_wots160_sig, byte_array_to_wots256_sig, checksig_verify_to_limbs, wots160_sig_to_byte_array, wots256_sig_to_byte_array, WOTSPubKey}}, execute_script, signatures::wots_api::{wots160, wots256, SignatureImpl}};
+    use crate::{bn254::{fp254impl::Fp254Impl, fq::Fq}, chunk::{elements::CompressedStateObject, helpers::extern_hash_fps, wrap_wots::{byte_array_to_wots160_sig, byte_array_to_wots256_sig, checksig_verify_to_limbs, wots160_sig_to_byte_array, wots256_sig_to_byte_array, WOTSPubKey}}, execute_script, signatures::wots_api::{wots160, wots256, SignatureImpl}};
 
     #[test]
     fn test_wots256_sig_to_byte_array() {
